backend/requirements.txt
# Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0

# LangGraph / LLM
langgraph>=0.2.53
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
openai>=1.50.0
anthropic>=0.34.0

# Utils
python-dotenv==1.0.1

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
aiosqlite==0.20.0
email-validator==2.2.0
